# openapi.yaml
openapi: 3.0.3
info:
  title: SafeGirl Backend API
  description: |
    Women's safety reporting platform with encrypted reports, blockchain verification,
    non-blocking panic alerts, and report sharing with access control.

    **Key Features:**
    - Phone-only OTP authentication (no passwords)
    - AES-256-GCM encryption for report data
    - IPFS + Polygon blockchain for immutable report records
    - Non-blocking panic alerts with SMS notifications
    - Report access control and sharing
    - Key recovery with PIN backup
  version: 1.0.0
  contact:
    name: SafeGirl Team
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.safegirl.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from signup/login verify endpoints
  schemas:
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [access_granted, panic_alert, report_submitted, report_updated, access_revoked]
        title:
          type: string
        message:
          type: string
        relatedId:
          type: string
          nullable: true
          description: reportId or accessId
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
    NotificationsResponse:
      type: object
      required: [notifications, total]
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    UnreadCountResponse:
      type: object
      required: [unreadCount]
      properties:
        unreadCount:
          type: integer
          example: 3

    # ========== KEY RECOVERY SCHEMAS ==========
...
    RecoverKeyRequest:
      type: object
      required: [phone, pinHash]
      properties:
        phone:
          type: string
          description: User's phone number
          example: "0750902921"
        pinHash:
          type: string
          description: SHA-256 hash of PIN
    RecoverKeyResponse:
      type: object
      required: [success, encryptedKey, keyIv, keyAuthTag]
      properties:
        success:
          type: boolean
          example: true
        encryptedKey:
          type: string
          description: Encrypted key (hex)
        keyIv:
          type: string
          description: IV (hex)
        keyAuthTag:
          type: string
          description: Auth tag (hex)

    # ========== SEARCH SCHEMAS ==========
...
paths:
...
  /api/notifications:
    get:
      summary: Get notifications for authenticated user
      operationId: getNotifications
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
  /api/notifications/unread/count:
    get:
      summary: Get unread notification count
      operationId: getUnreadCount
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
  /api/notifications/{notificationId}/read:
    post:
      summary: Mark notification as read
      operationId: markAsRead
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as read
  /api/notifications/mark-all-read:
    post:
      summary: Mark all notifications as read
      operationId: markAllAsRead
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All marked as read
  /api/notifications/{notificationId}:
    delete:
      summary: Delete notification
      operationId: deleteNotification
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification deleted

    post:
      summary: Recover encryption key using PIN and phone
      operationId: recoverKeyWithPin
      tags: [Key Recovery]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverKeyRequest'
      responses:
        '200':
          description: Key recovered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoverKeyResponse'
        '400':
          description: Invalid PIN
        '429':
          description: Rate limited (5 attempts per 15 minutes)

  # ========== SEARCH ENDPOINTS ==========
...